Precompiled script plugin or kotlin configuration
### What's done:
* Added kotlin-library.gradle.kts
* Use it in build.gradle.kts
