+++ build.gradle.kts
@@ -4,7 +4,6 @@ import org.cqfn.save.buildutils.configurePublishing
 import org.cqfn.save.buildutils.configureVersioning
 import org.cqfn.save.buildutils.createDetektTask
 import org.cqfn.save.buildutils.createDiktatTask
-import org.cqfn.save.buildutils.disableRedundantKmpTargets
 import org.cqfn.save.buildutils.installGitHooks
 
 plugins {
@@ -20,8 +19,6 @@ allprojects {
     }
     configureDiktat()
     configureDetekt()
-
-    disableRedundantKmpTargets()
 }
 createDiktatTask()
 createDetektTask()
